fix: allow a custom k8s ServiceAccount to be used
when using other clouds its often easier to configure the k8s ServiceAccount to be used.

e.g. on Google Cloud you can create a k8s SA that maps to an cloud SA and cloud role; but the 2 things need to be mapped together in terms of name + namespace; so its easier to create that and reuse it inside the Terraform pods.

fixes #6