allow the export runner to retry on failure
Sometimes export Runner fails when the target repo is updated after the runner has downloaded the repo and before it pushes changes. This results in the target being ahead of the export runner and the git command fails to push to the repo.  When a lot of TFO resources run at the same time, and they all want to use export repo, this tends to happen a lot. To fix this, allow the export runner to retry on failure. If true the pod restarts and the push attempt will be repeated.

Also allow the user to define the email and username of the automation user that pushes to git.
Configure as:

```yaml
kind: Terraform
spec:
    ...
    exportRepo:
        ...
        retryOnFailure: true
        gitUsername: automation user1
        gitEmail: autogit@company.com
```