Re-plan when apply fails
When tfo runs terraform apply, it uses a plan that was saved to the workflow persistent volume. If the apply fails, it makes the plan stale as it may not be relective of actual state anymore. In this case a new plan must be created.

So when an apply task fails, start the workflow from the planning stage again.