default secret namespace to the terraform resource
when parsing the SCM (git) secrets, must default to terraform's namespace. Even though the secretRef allows the defining of a namespace, TFO does not create a ClusterRole for the tasks and the workflow will fail anyway. But since the operator can reach out to other namespaces, in the case that a finalizer does need to be removed, the changes here will handle it.

Future dev (maybe me) please fix secret to same namespace as the terraform resource only.